requirements.txt
Flask==2.3.2
firebase-admin==6.2.0
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==20.1.0  # (For production)
python-jose==3.3.0  # (For JWT tokens)